wallet-api: Int instead Integer for value
* `newtype Value = Value Int` means we can use the same type for ada
  values PLC and Haskell, which allows us to get rid of a few
  `fromIntegral`s

+++ plutus-use-cases/test/Spec/Vesting.hs
@@ -118,7 +118,7 @@ canRetrieveFundsAtEnd = checkVestingTrace scen1 $ do
 
     -- Wallet 1 now has control of all the funds that were locked in the
     -- vesting scheme.
-    traverse_ (uncurry assertOwnFundsEq) [(w2, w2Funds), (w1, startingBalance + fromIntegral total)]
+    traverse_ (uncurry assertOwnFundsEq) [(w2, w2Funds), (w1, startingBalance + total)]
 
 -- | Vesting scenario with test parameters
 data VestingScenario = VestingScenario {
@@ -136,7 +136,7 @@ startingBalance = 1000
 -- | Amount of money left in wallet `Wallet 2` after committing funds to the
 --   vesting scheme
 w2Funds :: UTXO.Value
-w2Funds = startingBalance - fromIntegral total
+w2Funds = startingBalance - total
 
 -- | Total amount of money vested in the scheme `scen1`
 total :: Runtime.Value
